Refactoring
Renamed the com.kossyuzokwe.fantasy.jpa package to
com.kossyuzokwe.fantasy.dao
