--- src/main/java/com/kossyuzokwe/fantasy/annotation/UniqueUsernameValidator.java
@@ -1,27 +1,0 @@
-package com.kossyuzokwe.fantasy.annotation;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
-import com.kossyuzokwe.fantasy.jpa.UserRepository;
-
-public class UniqueUsernameValidator implements ConstraintValidator<UniqueUsername, String> {
-	
-	@Autowired
-	private UserRepository userRepository;
-
-	@Override
-	public void initialize(UniqueUsername constraintAnnotation) {
-	}
-
-	@Override
-	public boolean isValid(String username, ConstraintValidatorContext context) {
-		if (userRepository == null) {
-			return true;
-		}
-		return userRepository.findByUserName(username) == null;
-	}
-
-}
