+++ src/main/java/com/kossyuzokwe/fantasy/service/LeagueService.java
@@ -5,10 +5,10 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.kossyuzokwe.fantasy.jpa.LeagueRepository;
-import com.kossyuzokwe.fantasy.jpa.PlayerRepository;
-import com.kossyuzokwe.fantasy.jpa.TeamRepository;
-import com.kossyuzokwe.fantasy.jpa.UserRepository;
+import com.kossyuzokwe.fantasy.dao.LeagueRepository;
+import com.kossyuzokwe.fantasy.dao.PlayerRepository;
+import com.kossyuzokwe.fantasy.dao.TeamRepository;
+import com.kossyuzokwe.fantasy.dao.UserRepository;
 
 @Service
 public class LeagueService {
