+++ src/main/java/com/kossyuzokwe/fantasy/service/TeamService.java
@@ -9,10 +9,10 @@ import org.springframework.security.access.method.P;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
-import com.kossyuzokwe.fantasy.jpa.LeagueRepository;
-import com.kossyuzokwe.fantasy.jpa.PlayerRepository;
-import com.kossyuzokwe.fantasy.jpa.TeamRepository;
-import com.kossyuzokwe.fantasy.jpa.UserRepository;
+import com.kossyuzokwe.fantasy.dao.LeagueRepository;
+import com.kossyuzokwe.fantasy.dao.PlayerRepository;
+import com.kossyuzokwe.fantasy.dao.TeamRepository;
+import com.kossyuzokwe.fantasy.dao.UserRepository;
 import com.kossyuzokwe.fantasy.model.Team;
 import com.kossyuzokwe.fantasy.model.User;
