+++ src/main/webapp/WEB-INF/applicationContext.xml
@@ -19,7 +19,7 @@
 		<property name="entityManagerFactory" ref="emf"></property>
 	</bean>
 
-	<jpa:repositories base-package="com.kossyuzokwe.fantasy.jpa"
+	<jpa:repositories base-package="com.kossyuzokwe.fantasy.dao"
 		entity-manager-factory-ref="emf" />
 
 	<import resource="security.xml" />
